move lets config helpers to separate file
